Move `MissingMasterEmitter` to `Emitters` namespace, update references across `SkyrimSE` and `Fallout4`, and clean up unused folder declarations in the project file.
