+++ src/NexusMods.Games.CreationEngine/NexusMods.Games.CreationEngine.csproj
@@ -21,9 +21,5 @@
       <PackageReference Include="Mutagen.Bethesda.Fallout4" />
       <PackageReference Include="Mutagen.Bethesda.Skyrim" />
     </ItemGroup>
-    <ItemGroup>
-      <Folder Include="Emitters\" />
-      <Folder Include="SkyrimSE\Emitters\" />
-    </ItemGroup>
 
 </Project>
